added protocol buffer to describe a candidature
