+++ go.mod
@@ -4,5 +4,6 @@ go 1.14
 
 require (
 	github.com/gocarina/gocsv v0.0.0-20200330101823-46266ca37bd3
+	github.com/golang/protobuf v1.4.2
 	github.com/matryer/is v1.3.0
 )
